Wrapped the Audit log so the backend can be switched out
